Refactor the interval to aggregate actions
This is a significant change from my earlier proposal. In this
iteration, we use a stateful backing called `runner.js` which both
tracks which actions need to be executed and when. It coalesces
operations so that every action that needs to occur, for example, every
minute, does so together every minute instead of being scattered about.

Currently there are only a few intervals available _by design_ to keep
things simple and encourage aggregation. Further interval groups should
be added only with a proper reason why the existing set is insufficient.

The React component in this iteration is mostly an interface to the
underlying `runner.js`, but handles all of the visibility issues.
Further, the `leading` property has been removed because this interval
makes no guarantees about when exactly that the callbacks will execute,
only that they will do so on the specified interval. Probably the
default behavior ought to be that the callbacks get called initially
after being added to the list.