Interval: Ease on the return for Immutable.Map.forEach
When I wrote the code for the `forEach` handler I noticed a warning in
the documentation for Immutable.js that said any iteration returning
`false` would stop iteration. Because of this, I explicitly returned
true to make sure we went through the whole list.

Today I had a chance to test out the behavior and figure out exactly
what happens and so I have rewritten the code to make a better use of
this rule.

 - If the function inside of `forEach` returns === `false` then it will
   stop iteration. `undefined` and `null` do not terminate iteration.

 - Because we still pass of execution to `onTick()` which understandably
   could return `false`, I'm leaving in a guard around that call.